+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.example.wheretonext"
+        applicationId "com.alana.wheretonext"
         minSdk 29
         targetSdk 32
         versionCode 1
