+++ app/src/main/AndroidManifest.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.wheretonext">
+    package="com.alana.wheretonext">
 
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".ParseApplication"
+        android:name=".network.ParseApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -19,22 +19,22 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".FavoritePhrasesActivity"
+            android:name=".phrases.FavoritePhrasesActivity"
             android:exported="false" />
         <activity
-            android:name=".PhrasesActivity"
+            android:name=".phrases.PhrasesActivity"
             android:exported="false" />
         <activity
-            android:name=".MapActivity"
+            android:name=".map.MapActivity"
             android:exported="false" />
         <activity
-            android:name=".SignUpActivity"
+            android:name=".signup.SignUpActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".login.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
