+++ app/src/main/java/com/alana/wheretonext/login/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.wheretonext;
+package com.alana.wheretonext.login;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -11,9 +11,14 @@ import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.Toast;
 
+import com.alana.wheretonext.MainActivity;
+import com.alana.wheretonext.map.MapActivity;
+import com.alana.wheretonext.signup.SignUpActivity;
 import com.parse.LogInCallback;
 import com.parse.ParseException;
 import com.parse.ParseUser;
+
+import com.alana.wheretonext.R;
 
 public class LoginActivity extends AppCompatActivity {
