+++ app/src/main/java/com/alana/wheretonext/signup/SignUpActivity.java
@@ -1,4 +1,4 @@
-package com.example.wheretonext;
+package com.alana.wheretonext.signup;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -10,8 +10,11 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import com.alana.wheretonext.login.LoginActivity;
 import com.parse.ParseException;
 import com.parse.ParseUser;
+
+import com.alana.wheretonext.R;
 
 public class SignUpActivity extends AppCompatActivity {
